+++ src/file_pipe_log/pipe.rs
@@ -15,7 +15,9 @@ use crate::config::Config;
 use crate::env::FileSystem;
 use crate::event_listener::EventListener;
 use crate::metrics::*;
-use crate::pipe_log::{FileBlockHandle, FileId, FileSeq, LogFileContext, LogQueue, PipeLog};
+use crate::pipe_log::{
+    FileBlockHandle, FileId, FileSeq, LogFileContext, LogQueue, PipeLog, ReactiveBytes,
+};
 use crate::{perf_context, Error, Result};
 
 use super::format::{FileNameExt, LogFileFormat};
@@ -209,7 +211,6 @@ pub(super) struct SinglePipe<F: FileSystem> {
     queue: LogQueue,
     file_format: LogFileFormat,
     target_file_size: usize,
-    bytes_per_sync: usize,
     file_system: Arc<F>,
     listeners: Vec<Arc<dyn EventListener>>,
 
@@ -329,7 +330,6 @@ impl<F: FileSystem> SinglePipe<F> {
             queue,
             file_format: LogFileFormat::new(cfg.format_version, alignment),
             target_file_size: cfg.target_file_size.0 as usize,
-            bytes_per_sync: cfg.bytes_per_sync.0 as usize,
             file_system,
             listeners,
 
@@ -468,12 +468,24 @@ impl<F: FileSystem> SinglePipe<F> {
         reader.read(handle)
     }
 
-    fn append(&self, bytes: &[u8]) -> Result<FileBlockHandle> {
+    fn append<T: ReactiveBytes + ?Sized>(&self, bytes: &mut T) -> Result<FileBlockHandle> {
         fail_point!("file_pipe_log::append");
         let mut active_file = self.active_file.lock();
+        if active_file.writer.offset() >= self.target_file_size {
+            if let Err(e) = self.rotate_imp(&mut active_file) {
+                panic!(
+                    "error when rotate [{:?}:{}]: {}",
+                    self.queue, active_file.seq, e
+                );
+            }
+        }
+
         let seq = active_file.seq;
-        #[cfg(feature = "failpoints")]
         let format = active_file.format;
+        let ctx = LogFileContext {
+            id: FileId::new(self.queue, seq),
+            version: format.version,
+        };
         let writer = &mut active_file.writer;
 
         #[cfg(feature = "failpoints")]
@@ -497,7 +509,7 @@ impl<F: FileSystem> SinglePipe<F> {
             }
         }
         let start_offset = writer.offset();
-        if let Err(e) = writer.write(bytes, self.target_file_size) {
+        if let Err(e) = writer.write(bytes.as_bytes(&ctx), self.target_file_size) {
             if let Err(te) = writer.truncate() {
                 panic!(
                     "error when truncate {} after error: {}, get: {}",
@@ -548,15 +560,11 @@ impl<F: FileSystem> SinglePipe<F> {
         Ok(handle)
     }
 
-    fn maybe_sync(&self, force: bool) -> Result<()> {
+    fn sync(&self) -> Result<()> {
         let mut active_file = self.active_file.lock();
         let seq = active_file.seq;
         let writer = &mut active_file.writer;
-        if writer.offset() >= self.target_file_size {
-            if let Err(e) = self.rotate_imp(&mut active_file) {
-                panic!("error when rotate [{:?}:{}]: {}", self.queue, seq, e);
-            }
-        } else if writer.since_last_sync() >= self.bytes_per_sync || force {
+        {
             let _t = StopWatch::new(perf_context!(log_sync_duration));
             if let Err(e) = writer.sync() {
                 panic!("error when sync [{:?}:{}]: {}", self.queue, seq, e,);
@@ -622,14 +630,6 @@ impl<F: FileSystem> SinglePipe<F> {
         self.flush_metrics(current.total_len);
         Ok((current.first_seq_in_use - prev.first_seq_in_use) as usize)
     }
-
-    fn fetch_active_file(&self) -> LogFileContext {
-        let files = self.files.read();
-        LogFileContext {
-            id: FileId::new(self.queue, files.first_seq + files.fds.len() as u64 - 1),
-            version: files.fds.back().unwrap().format.version,
-        }
-    }
 }
 
 /// A [`PipeLog`] implementation that stores data in filesystem.
@@ -670,13 +670,17 @@ impl<F: FileSystem> PipeLog for DualPipes<F> {
     }
 
     #[inline]
-    fn append(&self, queue: LogQueue, bytes: &[u8]) -> Result<FileBlockHandle> {
+    fn append<T: ReactiveBytes + ?Sized>(
+        &self,
+        queue: LogQueue,
+        bytes: &mut T,
+    ) -> Result<FileBlockHandle> {
         self.pipes[queue as usize].append(bytes)
     }
 
     #[inline]
-    fn maybe_sync(&self, queue: LogQueue, force: bool) -> Result<()> {
-        self.pipes[queue as usize].maybe_sync(force)
+    fn sync(&self, queue: LogQueue) -> Result<()> {
+        self.pipes[queue as usize].sync()
     }
 
     #[inline]
@@ -697,11 +701,6 @@ impl<F: FileSystem> PipeLog for DualPipes<F> {
     #[inline]
     fn purge_to(&self, file_id: FileId) -> Result<usize> {
         self.pipes[file_id.queue as usize].purge_to(file_id.seq)
-    }
-
-    #[inline]
-    fn fetch_active_file(&self, queue: LogQueue) -> LogFileContext {
-        self.pipes[queue as usize].fetch_active_file()
     }
 }
 
@@ -768,7 +767,6 @@ mod tests {
         let cfg = Config {
             dir: path.to_owned(),
             target_file_size: ReadableSize::kb(1),
-            bytes_per_sync: ReadableSize::kb(32),
             ..Default::default()
         };
         let queue = LogQueue::Append;
@@ -780,17 +778,17 @@ mod tests {
 
         // generate file 1, 2, 3
         let content: Vec<u8> = vec![b'a'; 1024];
-        let file_handle = pipe_log.append(queue, &content).unwrap();
-        pipe_log.maybe_sync(queue, false).unwrap();
+        let file_handle = pipe_log.append(queue, &mut &content).unwrap();
         assert_eq!(file_handle.id.seq, 1);
+        assert_eq!(file_handle.offset, header_size);
+        assert_eq!(pipe_log.file_span(queue).1, 1);
+
+        let file_handle = pipe_log.append(queue, &mut &content).unwrap();
+        assert_eq!(file_handle.id.seq, 2);
         assert_eq!(file_handle.offset, header_size);
         assert_eq!(pipe_log.file_span(queue).1, 2);
 
-        let file_handle = pipe_log.append(queue, &content).unwrap();
-        pipe_log.maybe_sync(queue, false).unwrap();
-        assert_eq!(file_handle.id.seq, 2);
-        assert_eq!(file_handle.offset, header_size);
-        assert_eq!(pipe_log.file_span(queue).1, 3);
+        pipe_log.rotate(queue).unwrap();
 
         // purge file 1
         assert_eq!(pipe_log.purge_to(FileId { queue, seq: 2 }).unwrap(), 1);
@@ -801,13 +799,11 @@ mod tests {
 
         // append position
         let s_content = b"short content".to_vec();
-        let file_handle = pipe_log.append(queue, &s_content).unwrap();
-        pipe_log.maybe_sync(queue, false).unwrap();
+        let file_handle = pipe_log.append(queue, &mut &s_content).unwrap();
         assert_eq!(file_handle.id.seq, 3);
         assert_eq!(file_handle.offset, header_size);
 
-        let file_handle = pipe_log.append(queue, &s_content).unwrap();
-        pipe_log.maybe_sync(queue, false).unwrap();
+        let file_handle = pipe_log.append(queue, &mut &s_content).unwrap();
         assert_eq!(file_handle.id.seq, 3);
         assert_eq!(
             file_handle.offset,
@@ -833,11 +829,6 @@ mod tests {
         // leave only 1 file to truncate
         pipe_log.purge_to(FileId { queue, seq: 3 }).unwrap();
         assert_eq!(pipe_log.file_span(queue), (3, 3));
-
-        // fetch active file
-        let file_context = pipe_log.fetch_active_file(LogQueue::Append);
-        assert_eq!(file_context.version, cfg.format_version);
-        assert_eq!(file_context.id.seq, 3);
     }
 
     #[test]
@@ -925,7 +916,6 @@ mod tests {
         let cfg = Config {
             dir: path.to_owned(),
             target_file_size: ReadableSize(1),
-            bytes_per_sync: ReadableSize::kb(32),
             // super large capacity for recycling
             purge_threshold: ReadableSize::mb(100),
             enable_log_recycle: true,
@@ -942,9 +932,10 @@ mod tests {
         }
         let mut handles = Vec::new();
         for i in 0..10 {
-            handles.push(pipe_log.append(&content(i)).unwrap());
-            pipe_log.maybe_sync(true).unwrap();
+            handles.push(pipe_log.append(&mut &content(i)).unwrap());
+            pipe_log.sync().unwrap();
         }
+        pipe_log.rotate().unwrap();
         let (first, last) = pipe_log.file_span();
         assert_eq!(pipe_log.purge_to(last).unwrap() as u64, last - first);
         // Try to read stale file.
@@ -963,8 +954,8 @@ mod tests {
         // Try to reuse.
         let mut handles = Vec::new();
         for i in 0..10 {
-            handles.push(pipe_log.append(&content(i + 1)).unwrap());
-            pipe_log.maybe_sync(true).unwrap();
+            handles.push(pipe_log.append(&mut &content(i + 1)).unwrap());
+            pipe_log.sync().unwrap();
         }
         // Verify the data.
         for (i, handle) in handles.into_iter().enumerate() {