+++ tests/failpoints/test_io_error.rs
@@ -84,7 +84,6 @@ fn test_file_write_error() {
         .unwrap();
     let cfg = Config {
         dir: dir.path().to_str().unwrap().to_owned(),
-        bytes_per_sync: ReadableSize::kb(1024),
         target_file_size: ReadableSize::kb(1024),
         ..Default::default()
     };
@@ -133,7 +132,6 @@ fn test_file_rotate_error() {
         .unwrap();
     let cfg = Config {
         dir: dir.path().to_str().unwrap().to_owned(),
-        bytes_per_sync: ReadableSize::kb(1024),
         target_file_size: ReadableSize::kb(4),
         ..Default::default()
     };
@@ -149,6 +147,9 @@ fn test_file_rotate_error() {
         .unwrap();
     engine
         .write(&mut generate_batch(1, 3, 4, Some(&entry)), false)
+        .unwrap();
+    engine
+        .write(&mut generate_batch(1, 4, 5, Some(&entry)), false)
         .unwrap();
     assert_eq!(engine.file_span(LogQueue::Append).1, 1);
     // The next write will be followed by a rotate.
@@ -209,7 +210,6 @@ fn test_concurrent_write_error() {
         .unwrap();
     let cfg = Config {
         dir: dir.path().to_str().unwrap().to_owned(),
-        bytes_per_sync: ReadableSize::kb(1024),
         target_file_size: ReadableSize::kb(1024),
         ..Default::default()
     };
@@ -380,11 +380,7 @@ fn test_no_space_write_error() {
             // Disk is `full` -> `spare`, the first `write` operation should failed.
             let _f1 = FailGuard::new("file_pipe_log::force_no_free_space", "1*return->off");
             let _f2 = FailGuard::new("log_fd::write::no_space_err", "1*return->off");
-            let cfg_err = Config {
-                target_file_size: ReadableSize(1),
-                ..cfg.clone()
-            };
-            let engine = Engine::open(cfg_err).unwrap();
+            let engine = Engine::open(cfg.clone()).unwrap();
             engine
                 .write(&mut generate_batch(4, 11, 21, Some(&entry)), true)
                 .unwrap_err();
@@ -456,7 +452,6 @@ fn test_non_atomic_write_error() {
         .unwrap();
     let cfg = Config {
         dir: dir.path().to_str().unwrap().to_owned(),
-        bytes_per_sync: ReadableSize::kb(1024),
         target_file_size: ReadableSize::kb(1024),
         ..Default::default()
     };